Enableing local mode via -l

+++ bin/index_db.sh
@@ -24,12 +24,15 @@ usage ()
 opennlp="None"
 eval="false"
 data_only="false"
+local_mode="false"
 
-while getopts "edo:" opt; do
+
+while getopts "ledo:" opt; do
   case $opt in
     o) opennlp="$OPTARG";;
     e) eval="true";;
     d) data_only="true"
+    l) local_mode="true"
   esac
 done
 
@@ -133,32 +136,73 @@ fi
 # Stop processing if one step fails
 set -e
 
-#Load the dump into HDFS:
-if hadoop fs -test -e ${LANGUAGE}wiki-latest-pages-articles.xml ; then
-  echo "Dump already in HDFS."
-else
-  echo "Loading Wikipedia dump into HDFS..."
-  if [ "$eval" == "false" ]; then
-      curl -# "http://dumps.wikimedia.org/${LANGUAGE}wiki/latest/${LANGUAGE}wiki-latest-pages-articles.xml.bz2" | bzcat | hadoop fs -put - ${LANGUAGE}wiki-latest-pages-articles.xml
-  else
-      curl -# "http://dumps.wikimedia.org/${LANGUAGE}wiki/latest/${LANGUAGE}wiki-latest-pages-articles.xml.bz2" | bzcat | python $BASE_WDIR/pig/pignlproc/utilities/split_train_test.py 12000 $WDIR/heldout.txt | hadoop fs -put - ${LANGUAGE}wiki-latest-pages-articles.xml
+if [ "$local_mode" == "true" ]; then
+
+  if [ ! -e "$BASE_WDIR/pig/pig-0.10.1/" ]; then
+    #Install pig:
+    cd $BASE_WDIR/pig
+    wget http://apache.mirror.triple-it.nl/pig/pig-0.10.1/pig-0.10.1-src.tar.gz
+    tar xvzf pig-0.10.1-src.tar.gz
+    rm pig-0.10.1-src.tar.gz
+    cd pig-0.10.1-src
+    ant jar
   fi
+
+  export PATH=$BASE_WDIR/pig/pig-0.10.1-src/bin:$PATH
+
+  #Get the dump
+  curl -# "http://dumps.wikimedia.org/${LANGUAGE}wiki/latest/${LANGUAGE}wiki-latest-pages-articles.xml.bz2" | bzcat > $WDIR/${LANGUAGE}wiki-latest-pages-articles.xml
+
+else
+  #Load the dump into HDFS:
+
+  if hadoop fs -test -e ${LANGUAGE}wiki-latest-pages-articles.xml ; then
+    echo "Dump already in HDFS."
+  else
+    echo "Loading Wikipedia dump into HDFS..."
+    if [ "$eval" == "false" ]; then
+        curl -# "http://dumps.wikimedia.org/${LANGUAGE}wiki/latest/${LANGUAGE}wiki-latest-pages-articles.xml.bz2" | bzcat | hadoop fs -put - ${LANGUAGE}wiki-latest-pages-articles.xml
+    else
+        curl -# "http://dumps.wikimedia.org/${LANGUAGE}wiki/latest/${LANGUAGE}wiki-latest-pages-articles.xml.bz2" | bzcat | python $BASE_WDIR/pig/pignlproc/utilities/split_train_test.py 12000 $WDIR/heldout.txt | hadoop fs -put - ${LANGUAGE}wiki-latest-pages-articles.xml
+    fi
+  fi
+
 fi
+
 
 
 #Load the stopwords into HDFS:
 echo "Moving stopwords into HDFS..."
 cd $BASE_DIR
-hadoop fs -put $STOPWORDS stopwords.$LANGUAGE.list || echo "stopwords already in HDFS"
 
 
-if [ -e "$opennlp/$LANGUAGE-token.bin" ]; then
-    hadoop fs -put "$opennlp/$LANGUAGE-token.bin" "$LANGUAGE.tokenizer_model" || echo "tokenizer model already in HDFS"
+
+
+if [ "$local_mode" == "false" ]; then
+
+  hadoop fs -put $STOPWORDS stopwords.$LANGUAGE.list || echo "stopwords already in HDFS"
+
+  if [ -e "$opennlp/$LANGUAGE-token.bin" ]; then
+      hadoop fs -put "$opennlp/$LANGUAGE-token.bin" "$LANGUAGE.tokenizer_model" || echo "tokenizer model already in HDFS"
+  else
+      touch empty;
+      hadoop fs -put empty "$LANGUAGE.tokenizer_model" || echo "tokenizer model already in HDFS"
+      rm empty;
+  fi
+
 else
-    touch empty;
-    hadoop fs -put empty "$LANGUAGE.tokenizer_model" || echo "tokenizer model already in HDFS"
-    rm empty;
+
+  cd $WDIR
+  cp $STOPWORDS stopwords.$LANGUAGE.list || echo "stopwords already in HDFS"
+
+  if [ -e "$opennlp/$LANGUAGE-token.bin" ]; then
+      cp "$opennlp/$LANGUAGE-token.bin" "$LANGUAGE.tokenizer_model" || echo "tokenizer already exists"
+  else
+      touch "$LANGUAGE.tokenizer_model"
+  fi
+
 fi
+
 
 #Adapt pig params:
 cd $BASE_DIR
@@ -166,22 +210,46 @@ cd $1/pig/pignlproc
 
 PIGNLPROC_JAR="$BASE_WDIR/pig/pignlproc/target/pignlproc-0.1.0-SNAPSHOT.jar"
 
+if [ "$local_mode" == "true" ]; then
+
+  mkdir -p $WDIR/pig_out/$LANGUAGE
+
+  PIG_INPUT="$WDIR/${LANGUAGE}wiki-latest-pages-articles.xml"
+  PIG_STOPWORDS="$WDIR/stopwords.$LANGUAGE.list"
+  TOKEN_OUTPUT="$WDIR/pig_out/$LANGUAGE/tokenCounts"
+  PIG_TEMPORARY_SFS="$WDIR/pig_out/$LANGUAGE/sf_lookup"
+  PIG_NE_OUTPUT="$WDIR/pig_out/$LANGUAGE/names_and_entities"
+
+  PIG_LOCAL="-x local"
+
+else
+
+  PIG_INPUT="/user/$USER/${LANGUAGE}wiki-latest-pages-articles.xml"
+  PIG_STOPWORDS="/user/$USER/stopwords.$LANGUAGE.list"
+  TOKEN_OUTPUT="/user/$USER/$LANGUAGE/tokenCounts"
+  PIG_TEMPORARY_SFS="/user/$USER/$LANGUAGE/sf_lookup"
+  PIG_NE_OUTPUT="/user/$USER/$LANGUAGE/names_and_entities"
+
+  PIG_LOCAL=""
+
+fi
+
 #Run pig:
-pig -param LANG="$LANGUAGE" \
+pig $PIG_LOCAL -param LANG="$LANGUAGE" \
     -param LOCALE="$2" \
-    -param INPUT="/user/$USER/${LANGUAGE}wiki-latest-pages-articles.xml" \
-    -param OUTPUT="/user/$USER/$LANGUAGE/names_and_entities" \
-    -param TEMPORARY_SF_LOCATION="/user/$USER/$LANGUAGE/sf_lookup" \
+    -param INPUT="$PIG_INPUT" \
+    -param OUTPUT="$PIG_NE_OUTPUT" \
+    -param TEMPORARY_SF_LOCATION="$PIG_TEMPORARY_SFS" \
     -param PIGNLPROC_JAR="$PIGNLPROC_JAR" \
     -param MACROS_DIR="$BASE_WDIR/pig/pignlproc/examples/macros/" \
     -m examples/indexing/names_and_entities.pig.params examples/indexing/names_and_entities.pig
 
 
-pig -param LANG="$LANGUAGE" \
+pig $PIG_LOCAL -param LANG="$LANGUAGE" \
     -param ANALYZER_NAME="$4Analyzer" \
-    -param INPUT="/user/$USER/${LANGUAGE}wiki-latest-pages-articles.xml" \
-    -param OUTPUT_DIR="/user/$USER/$LANGUAGE/tokenCounts" \
-    -param STOPLIST_PATH="/user/$USER/stopwords.$LANGUAGE.list" \
+    -param INPUT="$PIG_INPUT" \
+    -param OUTPUT_DIR="$TOKEN_OUTPUT" \
+    -param STOPLIST_PATH="$PIG_STOPWORDS" \
     -param STOPLIST_NAME="stopwords.$LANGUAGE.list" \
     -param PIGNLPROC_JAR="$PIGNLPROC_JAR" \
     -param MACROS_DIR="$BASE_WDIR/pig/pignlproc/examples/macros/" \
@@ -191,10 +259,23 @@ pig -param LANG="$LANGUAGE" \
 #Copy results to local:
 cd $BASE_DIR
 cd $WDIR
-hadoop fs -cat $LANGUAGE/tokenCounts/part* > tokenCounts
-hadoop fs -cat $LANGUAGE/names_and_entities/pairCounts/part* > pairCounts
-hadoop fs -cat $LANGUAGE/names_and_entities/uriCounts/part* > uriCounts
-hadoop fs -cat $LANGUAGE/names_and_entities/sfAndTotalCounts/part* > sfAndTotalCounts
+
+if [ "$local_mode" == "true" ]; then
+
+  cat $TOKEN_OUTPUT/part* > tokenCounts
+  cat $PIG_NE_OUTPUT/pairCounts/part* > pairCounts
+  cat $PIG_NE_OUTPUT/uriCounts/part* > uriCounts
+  cat $PIG_NE_OUTPUT/sfAndTotalCounts/part* > sfAndTotalCounts
+
+else
+
+  hadoop fs -cat $LANGUAGE/tokenCounts/part* > tokenCounts
+  hadoop fs -cat $LANGUAGE/names_and_entities/pairCounts/part* > pairCounts
+  hadoop fs -cat $LANGUAGE/names_and_entities/uriCounts/part* > uriCounts
+  hadoop fs -cat $LANGUAGE/names_and_entities/sfAndTotalCounts/part* > sfAndTotalCounts
+
+fi
+
 
 #Create the model:
 cd $BASE_DIR
